revert the k2 borrowing
