+++ nemo/collections/asr/parts/k2/graph_transducer.py
@@ -26,7 +26,7 @@ from nemo.core.utils.optional_libs import TRITON_AVAILABLE
 from nemo.utils import logging
 
 if TRITON_AVAILABLE:
-    from nemo.collections.asr.parts.triton.rnnt_logprobs import rnnt_logprobs_triton
+    from nemo.collections.asr.parts.k2.rnnt_logprobs_triton import rnnt_logprobs_triton
 
 
 def force_float32_context() -> ContextManager:
